Финальный проект 16 спринта - Решение проблемы с зависимостями

+++ gateway/src/main/java/ru/practicum/shareit/booking/BookingClient.java
@@ -50,7 +50,7 @@ public class BookingClient extends BaseClient {
         return get("?state={state}&from={from}&size={size}", userId, parameters);
     }
 
-    public ResponseEntity<Object> getByOwner(Integer userId, BookingState state, Long from, Long size) {
+    public ResponseEntity<Object> getByOwner(Integer userId, BookingState state, Integer from, Integer size) {
         Map<String, Object> parameters = Map.of(
                 "state", state.name(),
                 "from", from,
